separated hooks, cleaned up lint errors
