+++ frontend-builder/src/components/Builder/index.tsx
@@ -14,7 +14,6 @@ import Database from './Database'
 import Sidebar from './Sidebar'
 import Topbar from './Topbar'
 import FullScreenLoader from '../Shared/FullScreenLoader'
-import { useSelection } from '../../utils/hooks'
 import PropertyWindow from './Properties'
 
 const GET_PROJECT = gql`
@@ -33,7 +32,6 @@ const Builder: React.FC = function Builder() {
 
   const [canvasState, setCanvasState] =
     useState<CanvasState>(DEFAULT_CANVAS_STATE)
-  const { selection } = useSelection()
   const { data: projects } = useListProjectsQuery()
   const { data, loading } = useQuery(GET_PROJECT, { variables: { projectId } })
